Extra classes (interfaces) were removed. Storage classes implimented.

+++ src/dnekh/pp/carstore/carmodel/Hyundai.java
@@ -4,7 +4,7 @@ import dnekh.pp.carstore.bluprintclasses.Vehicle;
 
 public class Hyundai extends Vehicle {
 
-    public Hyundai(String id, String carClass, String carManufacturer, String carModel, int yearOfCarManufacturer, String color, boolean isUsed, int carPrice) {
+    public Hyundai(String id, String carClass, String carManufacturer, String carModel, int yearOfCarManufacturer, String color, boolean isUsed, double carPrice) {
         super(id, carClass, carManufacturer, carModel, yearOfCarManufacturer, color, isUsed, carPrice);
     }
 }
