+++ src/dnekh/pp/carstore/carmodel/Kia.java
@@ -4,7 +4,7 @@ import dnekh.pp.carstore.bluprintclasses.Vehicle;
 
 public class Kia extends Vehicle {
 
-    public Kia(String id, String carClass, String carManufacturer, String carModel, int yearOfCarManufacturer, String color, boolean isUsed, int carPrice) {
+    public Kia(String id, String carClass, String carManufacturer, String carModel, int yearOfCarManufacturer, String color, boolean isUsed, double carPrice) {
         super(id, carClass, carManufacturer, carModel, yearOfCarManufacturer, color, isUsed, carPrice);
     }
 }
