+++ src/dnekh/pp/carstore/carmodel/Toyota.java
@@ -5,7 +5,7 @@ import dnekh.pp.carstore.bluprintclasses.Vehicle;
 public class Toyota extends Vehicle {
 
 
-    public Toyota(String id, String carClass, String carManufacturer, String carModel, int yearOfCarManufacturer, String color, boolean isUsed, int carPrice) {
+    public Toyota(String id, String carClass, String carManufacturer, String carModel, int yearOfCarManufacturer, String color, boolean isUsed, double carPrice) {
         super(id, carClass, carManufacturer, carModel, yearOfCarManufacturer, color, isUsed, carPrice);
     }
 }
